+++ pages/api/events.js
@@ -1,5 +1,7 @@
 import logger from "@config/logger";
+import connectMongo from "@config/mongo";
 import Profile from "@models/Profile";
+import dateFormat from "@services/utils/dateFormat";
 
 export default async function handler(req, res) {
   if (req.method != "GET") {
@@ -13,44 +15,56 @@ export default async function handler(req, res) {
 }
 
 export async function getEvents() {
+  await connectMongo();
   let events = [];
   try {
     events = await Profile.aggregate([
       { $project: { username: 1, events: 1, isEnabled: 1 } },
       { $match: { "events.date.start": { $gt: new Date() }, isEnabled: true } },
       { $unwind: "$events" },
-      { $match: { "events.date.end": { $gt: new Date() } } },
-      { $sort: { "events.date.start": 1 } },
+      {
+        $match: {
+          "events.date.end": { $gt: new Date() },
+          "events.isEnabled": { $ne: false },
+        },
+      },
+      {
+        $group: {
+          _id: "$events.url",
+          usernames: { $addToSet: "$username" },
+          isVirtual: { $first: "$events.isVirtual" },
+          color: { $first: "$events.color" },
+          date: { $first: "$events.date" },
+          url: { $first: "$events.url" },
+          name: { $first: "$events.name" },
+          description: { $first: "$events.description" },
+          price: {$first: "$events.price"},
+          isEnabled: { $first: "$isEnabled" },
+        },
+      },
+      {
+        $sort: { "date.start": 1 },
+      },
     ]).exec();
 
-    events = events
-      .map((event) => ({
-        ...event.events,
-        username: event.username,
-        _id: event._id,
-      }))
-      // TODO remove and do with mongo query
-      .filter((event) => new Date(event.date.end) > new Date());
-
     let dateEvents = [];
-    const today = new Date()
+    const today = new Date();
     for (const event of events) {
       let cleanEvent = structuredClone(event);
-      const dateTimeStyle = {
-        dateStyle: "full",
-        timeStyle: "long",
-      };
       try {
-        cleanEvent.date.startFmt = new Intl.DateTimeFormat("en-GB", dateTimeStyle).format(
-          new Date(event.date.start)
-        );
-        cleanEvent.date.endFmt = new Intl.DateTimeFormat("en-GB", dateTimeStyle).format(
-          new Date(event.date.end)
-        );
+        cleanEvent.date.startFmt = dateFormat({
+          format: "long",
+          date: event.date.start,
+        });
+        cleanEvent.date.endFmt = dateFormat({
+          format: "long",
+          date: event.date.end,
+        });
 
-        cleanEvent.date.cfpOpen = event.date.cfpClose && (new Date(event.date.cfpClose) > today);
-  
-        dateEvents.push(cleanEvent)
+        cleanEvent.date.cfpOpen =
+          event.date.cfpClose && new Date(event.date.cfpClose) > today;
+
+        dateEvents.push(cleanEvent);
       } catch (e) {
         logger.error(e, `ERROR event date for: "${event.name}"`);
       }