+++ pages/events.js
@@ -11,7 +11,7 @@ import { EventTabs } from "@components/event/EventTabs";
 import PageHead from "@components/PageHead";
 import Badge from "@components/Badge";
 
-export async function getServerSideProps(context) {
+export async function getServerSideProps() {
   let events = await getEvents();
 
   return {
@@ -86,11 +86,11 @@ export default function Events({ events }) {
         <div className="flex flex-row items-center">
           <Badge
             content="?"
-            path="/docs/how-to-guides/events"
+            path="/docs/how-to-guides/events-forms"
             title="Go To Event Docs"
             badgeClassName={"translate-x-2/4 -translate-y-1/2"}
           >
-            <h1 className="text-4xl mb-4 font-bold ">Community events</h1>
+            <h1 className="text-4xl mb-4 font-bold ">Community Events</h1>
           </Badge>
         </div>
         <EventTabs
@@ -101,11 +101,11 @@ export default function Events({ events }) {
         <h2 className="text-md md:text-2xl text-lg text-primary-high font-bold md:mb-6 mb-3">
           {tabFilters.find((filter) => filter.key === eventType).description}
         </h2>
-        <ul role="list" className="divide-y divide-primary-low mt-6">
+        <ul role="list" className="mt-6">
           {categorizedEvents[eventType]?.map((event) => (
             <EventCard
               event={event}
-              username={event.username}
+              usernames={event.usernames}
               key={`${event.name} ${event.username}`}
             />
           ))}