+++ pages/map.js
@@ -2,13 +2,13 @@ import { useState } from "react";
 import dynamic from "next/dynamic";
 
 import logger from "@config/logger";
-import Tag from "@components/Tag";
+import Tag from "@components/tag/Tag";
 import Button from "@components/Button";
 import PageHead from "@components/PageHead";
 import Page from "@components/Page";
 import Badge from "@components/Badge";
 import { getTags } from "./api/discover/tags";
-import { getUsers } from "./api/profiles";
+import { getProfiles } from "./api/profiles";
 import config from "@config/app.json";
 
 //this is required as leaflet is not compatible with SSR
@@ -23,7 +23,7 @@ export async function getStaticProps() {
     tags: [],
   };
   try {
-    data.users = await getUsers();
+    data.users = await getProfiles();
   } catch (e) {
     logger.error(e, "ERROR search users");
   }
@@ -36,6 +36,45 @@ export async function getStaticProps() {
       user.location.name !== "unknown" &&
       user.location.provided.toLowerCase() !== "remote"
   );
+
+  // Apply offset equally to 4 quadrants arround point
+  const adjustCoords = (coords, offset, offset2, index) => {
+    switch (index % 4) {
+      case 0:
+        return [coords[0] + offset, coords[1] + offset2];
+      case 1:
+        return [coords[0] - offset, coords[1] + offset2];
+      case 2:
+        return [coords[0] - offset, coords[1] - offset2];
+      default:
+        return [coords[0] + offset, coords[1] - offset2];
+    }
+  };
+
+  data.users = data.users.map((user, index) => {
+    const offset = Math.random() * 0.02; // ~2.2km
+    const offset2 = Math.random() * 0.02; // ~2.2km
+    return {
+      type: "Feature",
+      properties: {
+        cluster: false,
+        tags: user.tags || [],
+        username: user.username,
+        name: user.name,
+        location: user.location.provided,
+        bio: user.bio || "",
+      },
+      geometry: {
+        type: "Point",
+        coordinates: adjustCoords(
+          [parseFloat(user.location.lon), parseFloat(user.location.lat)],
+          offset,
+          offset2,
+          index
+        ),
+      },
+    };
+  });
 
   try {
     data.tags = await getTags(true);
@@ -74,11 +113,11 @@ export default function Map({ data }) {
     const terms = [...updateSelectedTagsFilter(value)];
 
     results = users.filter((user) => {
-      if (user.name.toLowerCase().includes(valueLower)) {
+      if (user.properties.name.toLowerCase().includes(valueLower)) {
         return true;
       }
 
-      let userTags = user.tags?.map((tag) => tag.toLowerCase());
+      let userTags = user.properties.tags?.map((tag) => tag.toLowerCase());
 
       if (terms.every((keyword) => userTags?.includes(keyword.toLowerCase()))) {
         return true;
@@ -132,12 +171,15 @@ export default function Map({ data }) {
             content={
               filteredUsers.length > 0 ? filteredUsers.length : users.length
             }
+            badgeClassName={"translate-x-3 -translate-y-3"}
           >
             <Button
               onClick={resetFilter}
               primary={true}
               disable={selectedTags.size == 0 ? true : false}
-            >Clear/Reset Filters</Button>
+            >
+              Clear/Reset Filters
+            </Button>
           </Badge>
           {tags &&
             tags