+++ pages/playground.js
@@ -7,9 +7,19 @@ import Input from "@components/form/Input";
 import UserPage from "@components/user/UserPage";
 import Notification from "@components/Notification";
 import { clientEnv } from "@config/schemas/clientSchema";
+import {
+  profileSchema,
+  usernameSchema,
+} from "@config/schemas/jsonProfileSchemas";
+import * as z from "zod";
 
-export default function Playground() {
-  const BASE_URL = clientEnv.NEXT_PUBLIC_BASE_URL;
+export async function getServerSideProps() {
+  return {
+    props: { BASE_URL: clientEnv.NEXT_PUBLIC_BASE_URL },
+  };
+}
+
+export default function Playground({ BASE_URL }) {
   const defaultJson = `{
     "name": "Your Name",
     "bio": "Write a short bio about yourself",
@@ -25,6 +35,7 @@ export default function Playground() {
   const [validateComplete, setValidateComplete] = useState(false);
   const [formatComplete, setFormatComplete] = useState(false);
   const [errorMessage, setErrMsg] = useState("");
+  const [errors, setErrors] = useState([]);
   const [successMessage, setSuccessMsg] = useState("");
   const [gitUsername, setGitUsername] = useState("");
   const [previewModalState, setPreviewModalState] = useState(false);
@@ -32,8 +43,10 @@ export default function Playground() {
   const [showNotification, setShowNotification] = useState(false);
 
   const handleValidateJson = () => {
+    setErrors([]);
     try {
-      JSON.parse(profileJson);
+      const parsedProfile = JSON.parse(profileJson);
+      profileSchema.parse(parsedProfile);
       setSuccessMsg("Valid Json");
       setErrMsg("");
       setValidateComplete(true);
@@ -41,8 +54,12 @@ export default function Playground() {
       setTimeout(() => setShowNotification(false), 1500);
       return true;
     } catch (err) {
-      setErrMsg(err.toString());
-      setError(true);
+      if (err instanceof z.ZodError || z.ZodError.create(err).errors) {
+        setErrMsg("Invalid JSON. Please check the errors below.");
+        setErrors(err.errors);
+      } else {
+        setErrMsg(err.toString());
+      }
       setSuccessMsg("");
       setShowNotification(true);
       setTimeout(() => setShowNotification(false), 1500);
@@ -66,12 +83,20 @@ export default function Playground() {
   };
 
   const handlePreview = () => {
-    if (!gitUsername) {
-      setErrMsg("Github username required");
+    setErrors([]);
+    try {
+      usernameSchema.parse(gitUsername);
+    } catch (e) {
+      const errors = [...e.errors];
+      errors[0].path = ["GitHub username"];
+      setErrMsg("GitHub username required");
+      setErrors(e.errors);
       setSuccessMsg("");
       setShowNotification(true);
       setTimeout(() => setShowNotification(false), 1500);
+      return false;
     }
+
     try {
       if (gitUsername && profileJson && handleValidateJson()) {
         setErrMsg("");
@@ -81,8 +106,8 @@ export default function Playground() {
         setPreviewModalData(actualJson);
         setPreviewModalState(true);
       }
-    } catch (err) {
-      setErrMsg(err.toString());
+    } catch (e) {
+      setErrMsg(e.toString());
       setSuccessMsg("");
       setShowNotification(true);
       setTimeout(() => setShowNotification(false), 1500);
@@ -149,7 +174,7 @@ export default function Playground() {
           Your Profile Json
         </label>
         <textarea
-          className="h-80 dark:bg-primary-high dark:text-white border-2 hover:border-tertiary-medium focus:ring-0 focus:border-tertiary-medium focus:outline-0 transition-all duration-250 ease-linear rounded px-6 py-2 mb-4 block w-full"
+          className="h-80 dark:bg-primary-high dark:text-primary-low border-2 hover:border-tertiary-medium focus:ring-0 focus:border-tertiary-medium focus:outline-0 transition-all duration-250 ease-linear rounded px-6 py-2 mb-4 block w-full"
           id="profileJson"
           name="profileJson"
           value={profileJson}
@@ -161,6 +186,20 @@ export default function Playground() {
             setErrMsg("");
           }}
         />
+
+        <span className="mb-5">
+          {errors.map((error, idx) => {
+            return (
+              <p key={idx} className="text-red-500">
+                {error.path.length === 1
+                  ? error.path
+                  : `${error.path[0]} - ${error.path[2]}`}
+                : {error.message}
+              </p>
+            );
+          })}
+        </span>
+
         <div className="flex flex-row justify-end mb-3 gap-2">
           <Button {...buttonProps()} />
         </div>