+++ tests/events.spec.js
@@ -1,6 +1,6 @@
 // @ts-check
 import { test, expect } from "@playwright/test";
-const AxeBuilder = require("@axe-core/playwright").default;
+import AxeBuilder from "@axe-core/playwright";
 
 test("Click on events profile in navbar navigates to events page", async ({
   page,
@@ -10,7 +10,12 @@ test("Click on events profile in navbar navigates to events page", async ({
     .getByRole("navigation")
     .getByRole("link", { name: "Events" })
     .click();
-  await expect(page).toHaveURL("/events");
+  await expect(page).toHaveURL(/\/events/);
+});
+
+test("Events has title", async ({ page }) => {
+  await page.goto("/events");
+  await expect(page.locator("h1")).toHaveText("Community Events");
 });
 
 test.fixme("Events listed", async ({ page }) => {