+++ CHANGELOG.md
@@ -1,4 +1,45 @@
-## 1.141.1 (2023-06-11)
+## [1.173.1](https://github.com/EddieHubCommunity/LinkFree/compare/v1.173.0...v1.173.1) (2023-08-13)
+
+
+### Bug Fixes
+
+* center 404 image ([#8623](https://github.com/EddieHubCommunity/LinkFree/issues/8623)) ([6a710b7](https://github.com/EddieHubCommunity/LinkFree/commit/6a710b7448ad3f8a8872c4a84eaebe6e3b74d98c))
+
+
+
+# [1.173.0](https://github.com/EddieHubCommunity/LinkFree/compare/v1.172.3...v1.173.0) (2023-08-12)
+
+
+### Features
+
+* discover page for recently updated profiles ([#8615](https://github.com/EddieHubCommunity/LinkFree/issues/8615)) ([d8f43f9](https://github.com/EddieHubCommunity/LinkFree/commit/d8f43f9fc8d835335f715f1a1e886068d805b1b1))
+
+
+
+## [1.172.3](https://github.com/EddieHubCommunity/LinkFree/compare/v1.172.2...v1.172.3) (2023-08-10)
+
+
+### Bug Fixes
+
+* users search results shows random first ([#8514](https://github.com/EddieHubCommunity/LinkFree/issues/8514)) ([e5be05c](https://github.com/EddieHubCommunity/LinkFree/commit/e5be05c804032b05dab05f042138e4f9e0554010))
+
+
+
+## [1.172.2](https://github.com/EddieHubCommunity/LinkFree/compare/v1.172.1...v1.172.2) (2023-08-10)
+
+
+### Bug Fixes
+
+* testimonials shows zero if no testimonials exist for user ([#8314](https://github.com/EddieHubCommunity/LinkFree/issues/8314)) ([de25879](https://github.com/EddieHubCommunity/LinkFree/commit/de25879180765628f24a29ee169d88b1814828d1))
+
+
+
+## [1.172.1](https://github.com/EddieHubCommunity/LinkFree/compare/v1.172.0...v1.172.1) (2023-08-10)
+
+
+### Bug Fixes
+
+* modal close button fixed ([#8496](https://github.com/EddieHubCommunity/LinkFree/issues/8496)) ([4f473d3](https://github.com/EddieHubCommunity/LinkFree/commit/4f473d378ddd525b6d3dd6d6dc51b6b55677622f))